sudoku board
Near the end, there are 2 methods that end the game (endGame() and quitButtonActionPerformed())
Add the frames of the end screen to those
Also i threw in the jpg of the sudoku board. im not sure if it is able to run on other computers due to how i imported it
LMK if u need anything!